requirements.txt
# Core Simulation
pygame-ce
Pillow

# AI / ML
ultralytics
opencv-python
numpy
pandas
openpyxl

# Backend Server
fastapi
uvicorn[standard]
websockets

# Visualization (Streamlit app)
streamlit
plotly
cufflinks
